fix(configs): set config as configs path
